Add new Mandrel-based Graal native compiling build agent for Java proj
To build native images for Quarkus, a new agent is needed that uses
GraalVM instead of the typical JDK. In testing, Mandrel JDK does not
work in Alpine due to the usage of musl, with the compat layer not
fixing all of the failures. This image will also include nodejs/yarn/npm
for some secondary operations of the builds.